Ternary operators
Instead of using the wordy if else and verifying the gender based on in, changed it using enum to define male and female and then inialize as selectedGender; the in the set state instantiate the genders accordingly. at last apply the ternary operator to verify if the gender is male and then the color active or inactive.